front: move formatSignedDelta to utils
